+++ test/yaml_node_wrapper_test.dart
@@ -68,6 +68,10 @@ void main() {
     expect(map.nodes.keys, everyElement(TypeMatcher<YamlScalar>()));
     expect(map.nodes[YamlScalar.wrap('list')], equals([1, 2, 3]));
     expect(map.style, equals(CollectionStyle.ANY));
+    expect((map.nodes['list'] as YamlList).style, equals(CollectionStyle.ANY));
+    expect((map.nodes['map'] as YamlMap).style, equals(CollectionStyle.ANY));
+    expect((map['map'].nodes['nested'] as YamlList).style,
+        equals(CollectionStyle.ANY));
   });
 
   test('YamlMap.wrap() with a sourceUrl', () {
@@ -85,7 +89,6 @@ void main() {
     expect(map['list'].span, isNullSpan(source));
     expect(map['map'].span, isNullSpan(source));
     expect(map.nodes['scalar'].span, isNullSpan(source));
-    expect(map.style, equals(CollectionStyle.ANY));
   });
 
   test('YamlMap.wrap() with a sourceUrl and style', () {
@@ -99,6 +102,10 @@ void main() {
     }, sourceUrl: 'source', style: CollectionStyle.BLOCK);
 
     expect(map.style, equals(CollectionStyle.BLOCK));
+    expect((map.nodes['list'] as YamlList).style, equals(CollectionStyle.ANY));
+    expect((map.nodes['map'] as YamlMap).style, equals(CollectionStyle.ANY));
+    expect((map['map'].nodes['nested'] as YamlList).style,
+        equals(CollectionStyle.ANY));
   });
 
   test('YamlList.wrap() with no sourceUrl', () {
@@ -135,6 +142,9 @@ void main() {
     expect(list.nodes[2].span, isNullSpan(isNull));
     expect(list[2], 'value');
     expect(list.style, equals(CollectionStyle.ANY));
+    expect((list[0] as YamlList).style, equals(CollectionStyle.ANY));
+    expect((list[1] as YamlMap).style, equals(CollectionStyle.ANY));
+    expect((list[1]['nested'] as YamlList).style, equals(CollectionStyle.ANY));
   });
 
   test('YamlList.wrap() with a sourceUrl', () {
@@ -152,7 +162,6 @@ void main() {
     expect(list[0].span, isNullSpan(source));
     expect(list[1].span, isNullSpan(source));
     expect(list.nodes[2].span, isNullSpan(source));
-    expect(list.style, equals(CollectionStyle.ANY));
   });
 
   test('YamlList.wrap() with a sourceUrl and style', () {
@@ -166,6 +175,9 @@ void main() {
     ], sourceUrl: 'source', style: CollectionStyle.FLOW);
 
     expect(list.style, equals(CollectionStyle.FLOW));
+    expect((list[0] as YamlList).style, equals(CollectionStyle.ANY));
+    expect((list[1] as YamlMap).style, equals(CollectionStyle.ANY));
+    expect((list[1]['nested'] as YamlList).style, equals(CollectionStyle.ANY));
   });
 
   test('re-wrapped objects equal one another', () {